refactor(codegen): extract std::bus::__local_dispatch into stdlib/bus.rs (Round 1.24)
The only stdlib bus surface with a codegen impl path (the rest of
the bus codegen — publish/subscribe/wire — is Round 3 territory).
1 fn lifted into BusStdlib<'ctx>.

codegen.rs: 37801 -> 37733 (-68 lines).
